# backend/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "RecipeJoe.Api | v1",
    "version": "1.0.0"
  },
  "paths": {
    "/api/recipes/import": {
      "post": {
        "tags": [
          "ImportEndpoints"
        ],
        "operationId": "ImportRecipe",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecipeDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportProblem"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportProblem"
                }
              }
            }
          }
        }
      }
    },
    "/api/recipes/{id}": {
      "get": {
        "tags": [
          "RecipeJoe.Api"
        ],
        "operationId": "GetRecipe",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecipeDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/recipes/{id}/image": {
      "get": {
        "tags": [
          "RecipeJoe.Api"
        ],
        "operationId": "GetRecipeImage",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "image/jpeg": {
                "schema": {
                  "type": "string",
                  "format": "byte"
                }
              },
              "image/png": {
                "schema": {
                  "type": "string",
                  "format": "byte"
                }
              },
              "image/webp": {
                "schema": {
                  "type": "string",
                  "format": "byte"
                }
              },
              "image/gif": {
                "schema": {
                  "type": "string",
                  "format": "byte"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ImportFailure": {
        "enum": [
          "InvalidUrl",
          "Unreachable",
          "Blocked",
          "NotFound",
          "BadResponse",
          "ForbiddenAddress",
          "NoRecipe"
        ]
      },
      "ImportProblem": {
        "required": [
          "type",
          "title",
          "status",
          "kind"
        ],
        "type": "object",
        "properties": {
          "type": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "kind": {
            "$ref": "#/components/schemas/ImportFailure"
          }
        }
      },
      "ImportRequest": {
        "required": [
          "url"
        ],
        "type": "object",
        "properties": {
          "url": {
            "type": "string"
          }
        }
      },
      "RecipeDto": {
        "required": [
          "id",
          "title",
          "servings",
          "prepMinutes",
          "cookMinutes",
          "totalMinutes",
          "ingredientLines",
          "steps",
          "sourceUrl",
          "imageUrl",
          "createdAt"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string"
          },
          "servings": {
            "type": [
              "null",
              "string"
            ]
          },
          "prepMinutes": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "cookMinutes": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "totalMinutes": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "ingredientLines": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "steps": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "sourceUrl": {
            "type": "string"
          },
          "imageUrl": {
            "type": [
              "null",
              "string"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      }
    }
  },
  "tags": [
    {
      "name": "ImportEndpoints"
    },
    {
      "name": "RecipeJoe.Api"
    }
  ]
}